Echt curriculum ingeladen: 91 vakken (ISLTH, PMGV, MGV)
Bron: 'vakkenlijst update.xlsx' -> database/data/curriculum.csv, geladen door
CurriculumSeeder (idempotent, matcht op opleiding + code). Referentiedata,
geen persoonsgegevens. PABO volgt later.

Schema-aanpassingen die de lijst afdwong
- vakken.ec: smallint -> decimal(4,1). 28 vakken hebben 2,5 EC; als geheel
  getal zou dat stil naar beneden afronden en de jaartotalen 11 EC laten
  afwijken. Idem vaktoewijzingen.vrijstelling_ec. Alle (int)-casts op EC zijn
  vervangen; nieuwe helper App\Support\Ec::toon() toont "2,5" en "5".
- vakken.code: unique(code) -> unique(opleiding_id, code). Elf codes (o.a.
  B-QR02, B-KL01, B-SG01) bestaan in zowel de bachelor als de pre-master; het
  zijn aparte vakken met een eigen cijferlijst, docent en presentielijst.
- Nieuwe kolom vakken.keuzevak. De keuzeruimte wordt NIET automatisch
  toegewezen bij inschrijving; Studentenzaken kent haar per student toe.
  Bachelor jaar 4 is daardoor 40 EC verplicht + 55 EC keuzeruimte, in plaats
  van 95 EC verplicht. Overgangsbeoordeling telt alleen de keuzevakken mee die
  daadwerkelijk aan de inschrijving zijn toegewezen.
- blok mag leeg zijn: het vak loopt dan het hele studiejaar (stages,
  scripties, M-GV16a/b). Vakstructuur en vaktoewijzing itereerden hard over
  blok 1..4, waardoor die vakken onzichtbaar waren; nu tonen zij een aparte
  sectie "Hele studiejaar".

Beslissingen over de bronlijst
- Bij tegenstrijdigheid is de tekstkolom 'Blok' leidend: B-SC07 -> blok 4,
  B-AR06-15 -> blok 2.
- B-FQ02-B-FQ03 blijft een vak zonder vast blok (student kiest I of II).
- Elk vak krijgt een standaard toetsonderdeel 'Tentamen' (weging 100%), anders
  kan een docent er geen cijfer voor invoeren.

Synthetische vakken verwijderd
- De 9 voorbeeldvakken (ISLTH-*) stonden actief naast het echte curriculum:
  ISLTH jaar 1 telde 94 EC in plaats van 60, en elke nieuwe student kreeg ze
  automatisch toegewezen. Ze zijn definitief verwijderd, inclusief 90
  vaktoewijzingen, 15 resultaten, 94 presenties, 5 cijferlijsten en 3
  vrijstellingsbesluiten.
- Ze zijn verhuisd van ReferentieSeeder naar SynthetischVakSeeder: uitsluitend
  testfixture, bewust niet in DatabaseSeeder. De betrokken tests seeden die
  fixture nu expliciet.

Nog te doen: docenten koppelen aan de 91 vakken (docent_id is leeg, dus 'Mijn
vakken' is voor docenten leeg) en de toetsopbouw per vak verfijnen.

Beide PDF-handleidingen bijgewerkt en opnieuw gegenereerd. 202 tests groen
(13 nieuw in CurriculumTest).

+++ app/Http/Controllers/StudentController.php
@@ -120,7 +120,7 @@ class StudentController extends Controller
                     $cijferVakken->push([
                         'vak' => $vt->vak,
                         'eind' => ['status' => 'vr', 'cijfer' => null],
-                        'ec' => (int) $vt->vak->ec,
+                        'ec' => (float) $vt->vak->ec,
                     ]);
                 }
             }
